Add discover item type
