+++ website/projects/scimaps/src/app/pages/maps/maps.component.html
@@ -14,7 +14,7 @@
 
 <h2 class="header discover-maps">Discover Maps</h2>
 
-<sci-discover-listing class="discover-item" *ngFor="let item of displayItems" [discoverItem]="item" (openDrawer)="openDetails()"></sci-discover-listing>
+<sci-discover-listing class="discover-item" *ngFor="let item of displayItems" [discoverItem]="item" [type]="'maps'" (openDrawer)="openDetails()"></sci-discover-listing>
 
 <button *ngIf="moreMaps()" class="more-button" (click)="showMoreMaps()">Show More Maps</button>
 <div class="mobile-button" (click)="showModal = true;">
@@ -30,7 +30,7 @@
 
   <sci-carousel class="maps-carousel">
     <sci-carousel-item *ngFor="let discoverItem of discoverItems">
-      <sci-discover-listing [discoverItem]="discoverItem" (openDrawer)="openDetails()"></sci-discover-listing>
+      <sci-discover-listing [discoverItem]="discoverItem" [type]="'maps'" (openDrawer)="openDetails()"></sci-discover-listing>
     </sci-carousel-item>
   </sci-carousel>
 </div>
